pyproject.toml
[project]
name = "auth-service"
version = "0.1.0"
description = "Authentication service provider built with FastAPI"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi[standard]>=0.135.3",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.2.2",
    "supabase>=2.28.3",
]
